Rename: store 폴더 중복 제거
