+++ app/_components/header/_component/user-header.tsx
@@ -1,8 +1,8 @@
 import { useRef, useState } from 'react';
 import Image from 'next/image';
 import defaultProfileImg from 'public/assets/images/default-profile.png';
-import useUserStore from '@/store/useUserStore';
 
+import useUserStore from '@/_stores/useUserStore';
 import useBackgroundClick from '@/_hooks/useBackgroundClick';
 
 import HeaderDropdown from './header-dropdown';
@@ -30,14 +30,12 @@ export default function UserHeader() {
         <Image src={alarm} alt="alarm" />
       </div>
       <div className="ml-3 flex items-center justify-between gap-[10px] md:ml-[25px]" onClick={toggleDropdown} ref={dropdownRef}>
-        <div className="h-8 w-8 overflow-hidden rounded-full cursor-pointer">
-          <Image src={user?.user.profileImageUrl || defaultProfileImg} alt="profile image" width={32} height={32}/>
+        <div className="h-8 w-8 cursor-pointer overflow-hidden rounded-full">
+          <Image src={user?.user.profileImageUrl || defaultProfileImg} alt="profile image" width={32} height={32} />
         </div>
         <div className="relative flex items-center gap-1">
-          <div className="cursor-pointer">
-            {user?.user.nickname || 'user'}
-          </div>
-          <div className="h-4 w-4 cursor-pointer" >
+          <div className="cursor-pointer">{user?.user.nickname || 'user'}</div>
+          <div className="h-4 w-4 cursor-pointer">
             <Image src={ArrowDown} alt="dropdown arrow" className={`duration-500 ${showList ? 'rotate-180' : 'rotate-0'}`} />
           </div>
           {showList && <HeaderDropdown />}
